requirements/base.txt
-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
